add new FF and component
