+++ frontend/src/scenes/insights/InsightTooltip/LEGACY_InsightTooltip.scss
@@ -1,0 +1,81 @@
+@import '~/vars';
+
+.ph-graph-tooltip,
+.ant-popover {
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+    font-size: 14px;
+    overflow-x: hidden;
+    z-index: $z_graph_tooltip;
+    transition: all 0.4s;
+}
+
+.ph-graph-tooltip {
+    max-width: 480px;
+    padding: 8px 12px;
+    border: 1px solid $border;
+    border-radius: $radius;
+    background: #fafafa;
+}
+
+.inner-tooltip {
+    &.multiple {
+        header {
+            padding-bottom: 8px;
+            border-bottom: 1px solid $border;
+        }
+
+        footer {
+            margin-top: 0 !important;
+            padding-top: 6px;
+            border-top: 1px solid $border;
+        }
+    }
+
+    header {
+        font-weight: bold;
+        margin-bottom: $default_spacing / 2;
+    }
+
+    footer {
+        margin-top: 0.5rem;
+        font-weight: bold;
+        font-size: 85%;
+        color: $text_muted;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        svg {
+            font-size: 1.4em;
+            margin-right: 4px;
+        }
+    }
+
+    .series-container {
+        position: relative;
+    }
+
+    .inspect-persons-label {
+        display: flex;
+        align-items: center;
+        margin-top: 2px;
+        font-style: italic;
+        svg {
+            margin-left: 2px;
+            margin-right: 6px;
+        }
+    }
+
+    ul {
+        list-style-type: none;
+        padding: 0;
+        margin-bottom: 0;
+        max-height: 32vh;
+        overflow-y: hidden; // As the element has pointer-events: none to allow proper function, scrolling is not supported anyways
+
+        li {
+            display: flex;
+            align-items: center;
+            padding-bottom: 2px;
+        }
+    }
+}
