+++ frontend/src/scenes/insights/LineGraph/LEGACY_LineGraph.jsx
@@ -12,8 +12,11 @@ import { Annotations, annotationsLogic, AnnotationMarker } from 'lib/components/
 import { useEscapeKey } from 'lib/hooks/useEscapeKey'
 import './LineGraph.scss'
 import { InsightLabel } from 'lib/components/InsightLabel'
-import { InsightTooltip } from '../InsightTooltip/InsightTooltip'
+import { LEGACY_InsightTooltip } from '../InsightTooltip/LEGACY_InsightTooltip'
 import { dayjs } from 'lib/dayjs'
+import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
+import { FEATURE_FLAGS } from 'lib/constants'
+import { InsightTooltip } from 'scenes/insights/InsightTooltip/InsightTooltip'
 
 //--Chart Style Options--//
 Chart.defaults.global.legend.display = false
@@ -60,6 +63,7 @@ export function LEGACY_LineGraph({
     const { annotationsList, annotationsLoading } = !inSharedMode
         ? useValues(annotationsLogic({ insightId }))
         : { annotationsList: [], annotationsLoading: false }
+    const { featureFlags } = useValues(featureFlagLogic)
     const [leftExtent, setLeftExtent] = useState(0)
     const [boundaryInterval, setBoundaryInterval] = useState(0)
     const [topExtent, setTopExtent] = useState(0)
@@ -322,15 +326,19 @@ export function LEGACY_LineGraph({
 
                     ReactDOM.render(
                         <Provider store={getContext().store}>
-                            <InsightTooltip
-                                altTitle={altTitle}
-                                referenceDate={referenceDate}
-                                interval={interval}
-                                bodyLines={bodyLines}
-                                inspectPersonsLabel={onClick && showPersonsModal}
-                                preferAltTitle={tooltipPreferAltTitle}
-                                hideHeader={type === 'horizontalBar'}
-                            />
+                            {featureFlags[FEATURE_FLAGS.NEW_INSIGHT_TOOLTIPS] ? (
+                                <InsightTooltip />
+                            ) : (
+                                <LEGACY_InsightTooltip
+                                    altTitle={altTitle}
+                                    referenceDate={referenceDate}
+                                    interval={interval}
+                                    bodyLines={bodyLines}
+                                    inspectPersonsLabel={onClick && showPersonsModal}
+                                    preferAltTitle={tooltipPreferAltTitle}
+                                    hideHeader={type === 'horizontalBar'}
+                                />
+                            )}
                         </Provider>,
                         tooltipEl
                     )
